Restructure commands and app setup
