+++ src/app/cli/print_test.go
