+++ src/app/cli/total.go
