+++ src/app/cli/total_test.go
